Move key-value store out-of-tree
Moved the in-memory key-value store out of tree into the SDK.

The module appears like a general-purpose key-value store, so
it can benefit others, too. As part of this SDK, consumers
can use the store without necessarily having to use the rest
of SPIKE.

In addition, this will provide transparency about how secrets
are stored in memory, and will enable creating isolated test
suites independent of the SPIKE core itself.

Also, those who want to integrate with SPIKE, or want to use
a similar backing store (maybe as a backup medium, or caching
purposes) will be able to use the very same implementation
that SPIKE uses internally eliminating possible compatibility
concerns.

In addition, if the key-value store is used for external
purposes by the community, it will mean additional hardening
and battle-testing and potential to recover edge cases and
performance improvements that we might have overlooked.

Overall, it feels like a good idea.

Signed-off-by: Volkan Özçelik <me@volkan.io>